checkboxHack: Handle SPACE and ENTER key on button (label) to change checkbox state
Necessary when the label can be focused instead of the checkbox.
Disables browser pagedown action when SPACE is pressed on the label.
Caveat: these (most) scripts are loaded asynchronously.
On very slow connections a user might focus the checkbox and press SPACE before that happens.
In that case the page will scroll. This use-case is expected to be very rare,
therefore an accepted trade-off.

Bug: T254851
Change-Id: I07d5f59096323564e1e51247ea7d12eca90d8606