Skin: Fix deprecation of Skin::setupSkinUserCss()
Removal of the call from OutputPage broke many skins still overriding this method,
resulting in the skins' stylesheets not being loaded:
https://gerrit.wikimedia.org/r/c/mediawiki/core/+/619144/2/includes/OutputPage.php

The wfDeprecated() warning was ineffective as the method was not called.
Proper deprecation warning is provided using reflection
when the skin subclass overrides the method called.

The deprecation warning would be too early in the constructor
as the headers aren't sent yet when the class is instantiated,
therefore the warning is made by initPage().

Change-Id: I653532c6100a8bea5daf474f4cba2c39230f4a6b