Skin: Add comments to deprecation of Skin::setupSkinUserCss()
Removal of the call from OutputPage broke many skins still overriding this method,
resulting in the skins' stylesheets not being loaded, violating the Stable interface policy:
https://gerrit.wikimedia.org/r/c/mediawiki/core/+/619144/2/includes/OutputPage.php#b2985

The wfDeprecated() warning was ineffective as the method was not called.
This fix provides proper deprecation warning using reflection
if and only if the skin subclass overrides the method called.

The deprecation warning would be too early in the constructor
as the headers aren't sent yet when the class is instantiated,
therefore the warning is made by initPage().

Bug: T266735
Change-Id: I653532c6100a8bea5daf474f4cba2c39230f4a6b

# Conflicts:
#	includes/api/ApiParse.php
#	includes/skins/Skin.php